nothing of note

+++ src/test.cpp
@@ -23,7 +23,7 @@ if the last bit is 0
 int main(void){
     int8 hi = -128;
 
-    std::cout << hi;
+    std::cout << (int)hi;
     test();
